refactor: change hooks and validation in result filter
