+++ dot-lib/src/pages/home/modules/listProducts/styles.module.scss
@@ -1,6 +1,13 @@
+@import "../../../../styles//mixins";
+
 .listProducts{
   display: flex;
   flex-direction: column;
   gap: 1rem;
   padding: 1rem 0;
+
+  .filterNotResult{
+    @include flexCenter;
+    height: 100px;
+  }
 }
