--- dot-lib/src/pages/home/useHome.ts
@@ -1,14 +1,0 @@
-export type TypeSearchProps = 'product' | 'company'
-
-export interface GetProductsProps {
-  type: TypeSearchProps
-  text: string
-}
-
-export function useHome() {
-  async function getProducts({ type, text }: GetProductsProps) {
-    console.log('buscou ', type, text)
-  }
-
-  return { getProducts }
-}
